src/services/ocr/requirements.txt
python-dotenv>=0.19.0
agentic-doc>=0.1.0  # Replace with actual package name and version
beautifulsoup4>=4.9.3
pandas>=1.3.0 